backend/requirements.txt
Django==1.11.4
djangorestframework==3.6.4
pytz==2018.4
